requirement.txt
django
djangorestframework
GuardPyCaptcha
django-cors-headers
persiantools
